Fix potential issue on systems with 64-bit longs in RNG initialization.

+++ trunk/src/utils.c
@@ -557,7 +557,7 @@ initialize_mt(void)
                 "Couldn't read from /dev/urandom! Resorting to normal seeding method.");
     } else {
       do_rawlog(LT_ERR, "Seeded RNG from /dev/urandom");
-      init_by_array(buf, r / sizeof(unsigned long));
+      init_by_array(buf, r / sizeof(uint32_t));
       return;
     }
   } else
